Change project structure to use app directory (Expo router)
