+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "brixcolor",
   "version": "1.0.0",
-  "main": "node_modules/expo/AppEntry.js",
+  "main": "expo-router/entry",
   "scripts": {
     "start": "expo start",
     "android": "expo start --android",
@@ -15,7 +15,14 @@
     "react": "18.2.0",
     "react-native": "0.73.4",
     "typescript": "^5.3.0",
-    "@types/react": "~18.2.45"
+    "@types/react": "~18.2.45",
+    "expo-router": "~3.4.8",
+    "react-native-safe-area-context": "4.8.2",
+    "react-native-screens": "~3.29.0",
+    "expo-linking": "~6.2.2",
+    "expo-constants": "~15.4.5",
+    "react-native-web": "~0.19.6",
+    "react-dom": "18.2.0"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0"
